feat: Implemented Skeleton for Router
