+++ src/pages/About/About.tsx
@@ -1,0 +1,9 @@
+const About = () => {
+  return (
+    <>
+      <h2>About</h2>
+    </>
+  );
+};
+
+export default About;
