+++ src/pages/Header/Header.tsx
@@ -1,0 +1,9 @@
+const Header = () => {
+  return (
+    <>
+      <h2>Header</h2>
+    </>
+  );
+};
+
+export default Header;
