+++ src/pages/Page404/Page404.tsx
@@ -1,0 +1,9 @@
+const Page404 = () => {
+  return (
+    <>
+      <h2>404</h2>
+    </>
+  );
+};
+
+export default Page404;
